agentsociety/webapi/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
paho-mqtt>=2.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
ujson>=5.8.0 